requirements.txt
faust-streaming==0.10.4
